# Replace gob with straightforward map serializer