```
$ benchcmp before.bench after.bench
benchmark            old ns/op     new ns/op      delta
BenchmarkOpen-4      228703637     526006848      +129.99%
BenchmarkClose-4     278377942     1615689315     +480.39%
BenchmarkWrite-4     113152        109331         -3.38%
BenchmarkRead-4      2088          2110           +1.05%

benchmark            old allocs     new allocs     delta
BenchmarkOpen-4      471997         9645           -97.96%
BenchmarkClose-4     566237         9708           -98.29%
BenchmarkWrite-4     1              1              +0.00%
BenchmarkRead-4      1              1              +0.00%

benchmark            old bytes     new bytes     delta
BenchmarkOpen-4      48929955      41083932      -16.04%
BenchmarkClose-4     68193395      41080288      -39.76%
BenchmarkWrite-4     1280          1280          +0.00%
BenchmarkRead-4      16            16            +0.00%
```